Move validate_data into report_engine to break circular import.
